feat: complete module task
